move workaround to conftest for other test files
